Implement role-based authorization with permissions
- Added role and permission management to the application for fine-grained access control.
- To properly use the application, the user must create a super admin account via Tinker.
- The super admin account is required to assign roles and permissions to other users for proper access control.

+++ routes/api.php
@@ -16,21 +16,20 @@ Route::middleware('auth')->group(function () {
     Route::post('logout', [AuthController::class, 'logout']);
 
     //projects routes
-    Route::get('projects', [ProjectController::class,'index']);
-    Route::post('projects', [ProjectController::class,'store']);
-    Route::get('projects/{project}', [ProjectController::class,'show']);
-    Route::put('projects/{project}', [ProjectController::class,'update']);
-    Route::delete('projects/{project}', [ProjectController::class,'destroy']);
+    Route::get('projects', [ProjectController::class,'index'])->middleware(['permission:view project']);
+    Route::post('projects', [ProjectController::class,'store'])->middleware(['permission:create project']);
+    Route::get('projects/{project}', [ProjectController::class,'show'])->middleware(['permission:view project']);
+    Route::put('projects/{project}', [ProjectController::class,'update'])->middleware(['permission:update project']);;
+    Route::delete('projects/{project}', [ProjectController::class,'destroy'])->middleware(['permission:delete project']);
 
     //tasks routes
-    Route::get('projects/{project}/tasks', [TaskController::class,'index']);
-    Route::post('projects/{project}/tasks', [TaskController::class,'store']);
-    Route::get('tasks/{task}', [TaskController::class,'show']);
-    Route::put('tasks/{task}', [TaskController::class,'update']);
-    Route::delete('tasks/{task}', [TaskController::class,'destroy']);
-    Route::get('{user}/created-tasks', [TaskController::class,'tasksCreatedBy']);
-    Route::get('{user}/tasks', [TaskController::class,'tasksAssignedTo']);
-
+    Route::get('projects/{project}/tasks', [TaskController::class,'index'])->middleware(['permission:view task']);
+    Route::post('projects/{project}/tasks', [TaskController::class,'store'])->middleware(['permission:create task']);
+    Route::get('tasks/{task}', [TaskController::class,'show'])->middleware(['permission:view task']);
+    Route::put('tasks/{task}', [TaskController::class,'update'])->middleware(['permission:update task']);
+    Route::delete('tasks/{task}', [TaskController::class,'destroy'])->middleware(['permission:delete task']);
+    Route::get('{user}/created-tasks', [TaskController::class,'tasksCreatedBy'])->middleware(['permission:view task']);
+    Route::get('{user}/tasks', [TaskController::class,'tasksAssignedTo'])->middleware(['permission:view task']);
 
 });
